pyproject.toml
[project]
name = "dnslookup"
version = "0.1.0"
description = ""
authors = [
    {name = "TheBiemGamer",email = "47662596+TheBiemGamer@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask (>=3.1.0,<4.0.0)",
    "dnspython (>=2.7.0,<3.0.0)",
    "flask-limiter (>=3.10.1,<4.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
